Move core to testable class
